refactor(File): FileHandler 일부 기능을 ImageUtil로 분리
