+++ src/main/java/page/clab/api/global/common/file/application/FileHandler.java
@@ -1,29 +1,19 @@
 package page.clab.api.global.common.file.application;
 
-import com.drew.imaging.ImageMetadataReader;
-import com.drew.imaging.ImageProcessingException;
-import com.drew.metadata.Directory;
-import com.drew.metadata.Metadata;
-import com.drew.metadata.MetadataException;
-import com.drew.metadata.exif.ExifIFD0Directory;
 import javax.imageio.ImageIO;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FilenameUtils;
-import org.imgscalr.Scalr;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 import page.clab.api.global.util.FileUtil;
-import page.clab.api.global.util.ImageCompressionUtil;
-import page.clab.api.global.util.TempFileUtil;
+import page.clab.api.global.util.ImageUtil;
 
 import java.awt.image.BufferedImage;
-import java.awt.image.BufferedImageOp;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Objects;
@@ -78,11 +68,11 @@ public class FileHandler {
         FileUtil.ensureParentDirectoryExists(file);
 
         try {
-            if (isImageFile(multipartFile)) {
-                BufferedImage originalImage = adjustImageDirection(multipartFile);
+            if (ImageUtil.isImageFile(multipartFile)) {
+                BufferedImage originalImage = ImageUtil.adjustImageDirection(multipartFile);
                 ImageIO.write(originalImage, Objects.requireNonNull(extension), file);
                 if (compressibleImageExtensions.contains(extension.toLowerCase())) {
-                    ImageCompressionUtil.compressImage(savePath, imageQuality);
+                    ImageUtil.compressImage(savePath, imageQuality);
                 }
             } else {
                 multipartFile.transferTo(file);
@@ -93,61 +83,6 @@ public class FileHandler {
 
         FileUtil.setFilePermissions(file, savePath);
         return savePath;
-    }
-
-    private BufferedImage adjustImageDirection(MultipartFile multipartFile) throws Exception {
-        Path tempFilePath = TempFileUtil.createSecureTempFile("temp", null);
-
-        File tempFile = tempFilePath.toFile();
-        multipartFile.transferTo(tempFile);
-
-        int originalDirection = getImageDirection(tempFile);
-        BufferedImage bufferedImage = ImageIO.read(tempFile);
-
-        switch (originalDirection) {
-            case 1:
-                break;
-            case 3:
-                bufferedImage = Scalr.rotate(bufferedImage, Scalr.Rotation.CW_180, (BufferedImageOp[]) null);
-                break;
-            case 6:
-                bufferedImage = Scalr.rotate(bufferedImage, Scalr.Rotation.CW_90, (BufferedImageOp[]) null);
-                break;
-            case 8:
-                bufferedImage = Scalr.rotate(bufferedImage, Scalr.Rotation.CW_270, (BufferedImageOp[]) null);
-                break;
-        }
-
-        if (tempFile.exists() && !tempFile.delete()) {
-            throw new IOException("Failed to delete image file: " + tempFile.getAbsolutePath());
-        }
-
-        return bufferedImage;
-    }
-
-    public int getImageDirection(File tempFile) {
-        int originalDirection = 1;
-        try {
-            Metadata metadata = ImageMetadataReader.readMetadata(tempFile);
-            Directory directory = metadata.getFirstDirectoryOfType(ExifIFD0Directory.class);
-            if (directory != null) {
-                originalDirection = directory.getInt(ExifIFD0Directory.TAG_ORIENTATION);
-            }
-        } catch (IOException e) {
-            log.error("이미지 파일을 읽는 중 IO 오류 발생: {}", e.getMessage());
-        } catch (ImageProcessingException e) {
-            log.error("이미지 파일 처리 중 오류 발생: {}", e.getMessage());
-        } catch (MetadataException e) {
-            log.error("이미지 파일의 메타데이터를 읽는 중 오류 발생: {}", e.getMessage());
-        } catch (Exception e) {
-            log.error("예기치 않은 오류 발생: {}", e.getMessage());
-        }
-        return originalDirection;
-    }
-
-    private boolean isImageFile(MultipartFile file) {
-        String contentType = file.getContentType();
-        return contentType != null && contentType.startsWith("image/");
     }
 
     public void deleteFile(String savedPath) {
